+++ examples/golf/sorted_integers.py
